refactor: Reservation 엔티티 패키지로 이동
