+++ src/main/java/roomescape/controller/RoomescapeController.java
@@ -9,8 +9,7 @@ import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RestController;
-import roomescape.Reservation;
-import roomescape.ReservationManager;
+import roomescape.entity.Reservation;
 import roomescape.ReservationCreateRequest;
 import roomescape.Reservations;
